requirements.txt
fastapi
uvicorn[standard]
transformers
torch
datasets
psycopg2-binary
scikit-learn
python-dotenv
openai
requests

